requirements.txt
boto3
joblib
numpy
pandas
scikit-learn
scipy
PyYAML

